Formatted the response for the hubcare api indicators
